Added DockerFile, lombok, test classes

+++ src/main/java/com/demo/vaccinebooking/Service/VaccineBookingService.java
@@ -1,5 +1,6 @@
 package com.demo.vaccinebooking.Service;
 
+import com.demo.vaccinebooking.Exception.VaccineBookingException;
 import com.demo.vaccinebooking.Model.APIResponse;
 import com.demo.vaccinebooking.Model.Booking;
 import com.demo.vaccinebooking.Model.Slot;
@@ -14,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.hibernate.annotations.common.util.impl.LoggerFactory;
 import org.jboss.logging.Logger;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 
 
 import javax.validation.ConstraintViolation;
@@ -22,6 +24,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class VaccineBookingService {
@@ -35,167 +38,116 @@ public class VaccineBookingService {
     SlotService slotService;
 
     //to get available slots for the given vaccine and date
-    public ResponseEntity<APIResponse> getAvailableSlots(String vaccineName, String date){
+    public ResponseEntity<APIResponse> getAvailableSlots(String vaccineName, String date) throws Exception{
         logger.info("get all available slots for a vaccine and date");
         APIResponse apiResponse = new APIResponse();
-        try {
-            LocalDate selectedDate = LocalDate.parse(date);
-            //get vaccine by name
-            Vaccine vaccine = vaccineService.getVaccineByName(vaccineName);
-            if (vaccine == null || vaccine.getAvailableShots() < 1) {
-                apiResponse.setSuccess(false);
-                apiResponse.setMessage("The vaccine you have given is unavailable. Please select a different vaccine");
-                return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.NOT_FOUND);
+        LocalDate selectedDate = LocalDate.parse(date);
+        //get vaccine by name
+        Optional<Vaccine> vaccine = vaccineService.getVaccineByName(vaccineName);
+        //to check the vaccine availability
+        if (!vaccine.isPresent() || vaccine.get().getAvailableShots() < 1) {
+            throw new VaccineBookingException("The vaccine you have given is unavailable. Please select a different vaccine", HttpStatus.NOT_FOUND);
+        } else {
+            //get available slots
+            List<Slot> slots = slotService.getSlotsByDate(selectedDate);
+            if (slots == null || slots.size() < 1) {
+                throw new VaccineBookingException("There are no slots available. Please try some other date.", HttpStatus.FORBIDDEN);
             } else {
-                //get available slots
-                List<Slot> slots = slotService.getSlotsByDate(selectedDate);
-                if (slots == null || slots.size() < 1) {
-                    apiResponse.setSuccess(false);
-                    apiResponse.setMessage("There are no slots available. Please try some other date.");
-                    return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.NOT_FOUND);
-                } else {
-                    apiResponse.setSuccess(true);
-                    apiResponse.setMessage("Here are the available slots for you.");
-                    List<Object> objectList = new ArrayList<>();
-                    for (Slot slot : slots) {
-                        if (slot.isAvailable()) {
-                            objectList.add(slot);
-                        }
-                    }
-                    //slots.stream().forEach(slot -> objectList.add(slot));
-                    apiResponse.setData(objectList);
-                    return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
-                }
-            }
-        }
-        catch (Exception e){
-            logger.info(e);
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("There is a technical error while retrieving the slots.");
-            List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.NOT_FOUND);
-        }
-    }
-
-    //to create booking
-    public ResponseEntity<APIResponse> createBooking(Booking booking)
-    {
-        logger.info("create a booking");
-        APIResponse apiResponse = new APIResponse();
-        int assingedSlotId = booking.getSlot().getSlotId();
-        try {
-            Slot assignedSlot = slotService.getSlotsById(assingedSlotId);
-            if(assignedSlot == null || assignedSlot.isAvailable() == false){
-                apiResponse.setSuccess(false);
-                apiResponse.setMessage("The slot you have selected is unavailable. Please choose a different time slot.");
-                return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
-            }
-
-            Vaccine vaccine = vaccineService.getVaccineByName(booking.getVaccineName());
-            if(vaccine == null || vaccine.getAvailableShots() < 1){
-                apiResponse.setSuccess(false);
-                apiResponse.setMessage("The vaccine you have selected is unavailable. Please choose a different vaccine.");
-                return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
-            }
-
-            //create booking
-            Booking createdBooking = bookingRepository.save(booking);
-            //assign slot to the booking
-            slotService.assignSlotToBooking(assingedSlotId);
-            //decrease available shots count
-            vaccineService.decreaseVaccineShots(booking.getVaccineName());
-            apiResponse.setSuccess(true);
-            apiResponse.setMessage("Your appointment has been booked successfully. Here is your booking id : "+createdBooking.getBookingId());
-            return  new ResponseEntity<APIResponse>(apiResponse,HttpStatus.OK);
-        }
-        catch (ConstraintViolationException e){
-            logger.info(e.getConstraintViolations());
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("There are some errors during the booking creation.");
-            List<Object> objectList = new ArrayList<>();
-            for(ConstraintViolation<?> ex : e.getConstraintViolations()){
-                logger.info(ex);
-                objectList.add(ex.getMessage());
-            }
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
-        }
-        catch (Exception e){
-            logger.info(e);
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("You booking cannot be created. Please see the error details below.");
-            List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
-        }
-    }
-
-    public Booking getBooking(int bookingId){
-        return bookingRepository.getById(bookingId);
-    }
-
-    //to get Booking
-    public ResponseEntity<APIResponse> getBookingById(int bookingId){
-        logger.info("Get Booking");
-        APIResponse apiResponse = new APIResponse();
-        try{
-            Booking booking = getBooking(bookingId);
-            if(booking == null || booking.getVaccineName() == null ){
-                apiResponse.setSuccess(false);
-                apiResponse.setMessage("No booking details available for this booking id.");
-                return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.NOT_FOUND);
-            }
-            else {
                 apiResponse.setSuccess(true);
-                apiResponse.setMessage("Here is your booking detail");
+                apiResponse.setMessage("Here are the available slots for you.");
                 List<Object> objectList = new ArrayList<>();
-                objectList.add(booking);
+                for (Slot slot : slots) {
+                    if (slot.isAvailable()) {
+                        objectList.add(slot);
+                    }
+                }
+                //slots.stream().forEach(slot -> objectList.add(slot));
                 apiResponse.setData(objectList);
                 return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
             }
         }
-        catch (Exception e) {
-            logger.info(e);
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("No booking details available for this booking id.");
+    }
+
+    //to create booking
+    public ResponseEntity<APIResponse> createBooking(Booking booking) throws Exception {
+        logger.info("create a booking");
+        APIResponse apiResponse = new APIResponse();
+
+        //get slot using slot id from the payload
+        int assingedSlotId = booking.getSlot().getSlotId();
+        Optional<Slot> assignedSlot = slotService.getSlotsById(assingedSlotId);
+        //check if the slot is not present or if it is unavailable
+        if(!assignedSlot.isPresent() || assignedSlot.get().isAvailable() == false){
+            throw new VaccineBookingException("The slot you have selected is unavailable. Please choose a different time slot.", HttpStatus.NOT_FOUND);
+        }
+
+        //get vaccine using vaccine name
+        Optional<Vaccine> vaccine = vaccineService.getVaccineByName(booking.getVaccineName());
+        //check if the vaccine is not available
+        if(!vaccine.isPresent() || vaccine.get().getAvailableShots() < 1){
+            throw new VaccineBookingException("The vaccine you have selected is unavailable. Please choose a different vaccine.", HttpStatus.NOT_FOUND);
+        }
+        logger.info("success - we can create booking");
+        //create booking
+        Booking createdBooking = bookingRepository.save(booking);
+        logger.info("assign slot");
+        //assign slot to the booking
+        slotService.assignSlotToBooking(assingedSlotId);
+        //decrease available shots count
+        vaccineService.decreaseVaccineShots(booking.getVaccineName());
+        apiResponse.setSuccess(true);
+        apiResponse.setMessage("Your appointment has been booked successfully. Here is your booking id : "+createdBooking.getBookingId());
+        return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.OK);
+
+    }
+
+    public Optional<Booking> getBooking(int bookingId){
+        return bookingRepository.findById(bookingId);
+    }
+
+    //to get Booking
+    public ResponseEntity<APIResponse> getBookingById(int bookingId) throws Exception {
+        logger.info("Get Booking");
+        APIResponse apiResponse = new APIResponse();
+        Optional<Booking> booking = getBooking(bookingId);
+        if(!booking.isPresent() || booking.get().getVaccineName() == null ){
+            throw new VaccineBookingException("No booking details available for this booking id.", HttpStatus.NOT_FOUND);
+        }
+        else {
+            apiResponse.setSuccess(true);
+            apiResponse.setMessage("Here is your booking detail");
             List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
+            objectList.add(booking);
             apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.NOT_FOUND);
+            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
         }
     }
 
     //to update booking
-    public ResponseEntity<APIResponse> updateBooking(Booking booking)
-    {
+    public ResponseEntity<APIResponse> updateBooking(Booking booking) throws Exception {
         logger.info("update booking");
         APIResponse apiResponse = new APIResponse();
-        try {
-            int bookingId = booking.getBookingId();
-            Booking oldBooking = getBooking(bookingId);
-
+        int bookingId = booking.getBookingId();
+        Optional<Booking> oldBooking = getBooking(bookingId);
+        if(!oldBooking.isPresent()){
+            throw new VaccineBookingException("No booking details available for this booking id.", HttpStatus.NOT_FOUND);
+        }
+        else {
             //check if vaccine is changed
             String vaccineName = booking.getVaccineName();
-            String oldVaccineName = oldBooking.getVaccineName();
-            if(!vaccineName.equals(oldVaccineName)){
-                apiResponse.setSuccess(false);
-                apiResponse.setMessage("Vaccine cannot be updated. Please cancel your booking and create a new one.");
-                return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.FORBIDDEN);
+            String oldVaccineName = oldBooking.get().getVaccineName();
+            if (!vaccineName.equals(oldVaccineName)) {
+                throw new VaccineBookingException("Vaccine cannot be updated. Please cancel your booking and create a new one.", HttpStatus.FORBIDDEN);
             }
 
-            int oldSlotId = oldBooking.getSlot().getSlotId();
+            int oldSlotId = oldBooking.get().getSlot().getSlotId();
             int newSlotId = booking.getSlot().getSlotId();
             //reassign slots if slotId changed
-            if(oldSlotId != newSlotId) {
+            if (oldSlotId != newSlotId) {
                 //check if the slot is available
-                Slot assignedSlot = slotService.getSlotsById(newSlotId);
-                if(assignedSlot == null || assignedSlot.isAvailable() == false){
-                    apiResponse.setSuccess(false);
-                    apiResponse.setMessage("The slot you have selected is unavailable. Please choose a different time slot.");
-                    return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
+                Optional<Slot> assignedSlot = slotService.getSlotsById(newSlotId);
+                if (!assignedSlot.isPresent() || assignedSlot.get().isAvailable() == false) {
+                    throw new VaccineBookingException("The slot you have selected is unavailable. Please choose a different time slot.", HttpStatus.NOT_FOUND);
                 }
                 slotService.unassignSlotFromBooking(oldSlotId);
                 slotService.assignSlotToBooking(booking.getSlot().getSlotId());
@@ -203,81 +155,41 @@ public class VaccineBookingService {
 
             apiResponse.setSuccess(true);
             apiResponse.setMessage("Your booking has been updated successfully.");
-            //List<Object> objectList = new ArrayList<>();
-            //objectList.add(bookingRepository.save(booking));
-            //apiResponse.setData(objectList);
             bookingRepository.save(booking);
             return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
         }
-        catch (ConstraintViolationException e){
-            logger.info(e.getConstraintViolations());
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("There are some errors during the booking creation.");
-            List<Object> objectList = new ArrayList<>();
-            for(ConstraintViolation<?> ex : e.getConstraintViolations()){
-                logger.info(ex);
-                objectList.add(ex.getMessage());
-            }
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.FORBIDDEN);
-        }
-        catch (Exception e){
-            logger.info(e);
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("Your booking cannot be updated.");
-            List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.FORBIDDEN);
-        }
     }
 
-    public ResponseEntity<APIResponse> deleteBooking(int bookingId){
+    public ResponseEntity<APIResponse> deleteBooking(int bookingId) throws Exception {
         logger.info("delete booking");
         APIResponse apiResponse = new APIResponse();
-        try {
-            //unassign slot from booking
-            Booking oldBooking = getBooking(bookingId);
-            int oldSlotId = oldBooking.getSlot().getSlotId();
+        //unassign slot from booking
+        Optional<Booking> oldBooking = getBooking(bookingId);
+        if(!oldBooking.isPresent()){
+            throw new VaccineBookingException("No booking details available for this booking id.", HttpStatus.NOT_FOUND);
+        }
+        else {
+            //unassign slot
+            int oldSlotId = oldBooking.get().getSlot().getSlotId();
             slotService.unassignSlotFromBooking(oldSlotId);
             //increase vaccine count
-            vaccineService.increaseVaccineShots(oldBooking.getVaccineName());
+            vaccineService.increaseVaccineShots(oldBooking.get().getVaccineName());
             //delete booking
             logger.info("delete booking");
             bookingRepository.deleteById(bookingId);
 
             apiResponse.setSuccess(true);
             apiResponse.setMessage("Your booking has been cancelled successfully.");
-            return new ResponseEntity<APIResponse>(apiResponse,HttpStatus.OK);
-        }
-        catch (Exception e){
-            logger.info(e);
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("There is a technical error while cancelling your booking.");
-            List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.FORBIDDEN);
+            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
         }
     }
 
     public ResponseEntity<APIResponse> deleteAllBookings(){
         logger.info("delete all bookings");
         APIResponse apiResponse = new APIResponse();
-        try {
-            bookingRepository.deleteAll();
-            //return "Bookings Deleted Successfully";
-            apiResponse.setSuccess(true);
-            apiResponse.setMessage("All bookings have been cancelled successfully.");
-            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
-        }
-        catch (Exception e){
-            apiResponse.setSuccess(false);
-            apiResponse.setMessage("There is a technical error while cancelling your booking.");
-            List<Object> objectList = new ArrayList<>();
-            objectList.add(e.getMessage());
-            apiResponse.setData(objectList);
-            return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.FORBIDDEN);
-        }
+        bookingRepository.deleteAll();
+        apiResponse.setSuccess(true);
+        apiResponse.setMessage("All bookings have been cancelled successfully.");
+        return new ResponseEntity<APIResponse>(apiResponse, HttpStatus.OK);
     }
 }
